create reproducible images
The images created by kernel2minor are not reproducible at the moment,
since the the uid, guid and file access/modification/change time is read
from the passed file.

Use fixed values instead to create the same header if a binary identical
file is used (which might have a different creating time).

Signed-off-by: Mathias Kresin <dev@kresin.me>